add list product function in user role and add role authentication

+++ routes/web.php
@@ -32,29 +32,28 @@ Route::post('/logout', [UserController::class, 'logoutUser'])->name('logout');
 Route::get('/contactUs', [UserController::class, 'contactUsView'])->name('contactUs');
 Route::post('/sendEmail', [UserController::class, 'sendEmail'])->name('send.email');
 
+// LIST PRODUCT
+Route::get('/listProduct', [UserController::class, 'listProductView'])->name('listProductView');
+
 // LANDING PAGE
 Route::get('/', [UserController::class, 'landingPageView']);
 
 // ----------------------------------------------ADMIN----------------------------------------------
 Route::prefix('/admin')->group(function () {
     // LIST PRODUCTS
-    Route::get('/listProduct', [AdminController::class, 'showProducts'])->middleware('auth')->name('listProduct');
+    Route::get('/listProduct', [AdminController::class, 'showProducts'])->middleware(['auth', 'role-auth'])->name('listProduct');
 
     // ADD PRODUCT
-    Route::get('/addProduct', [AdminController::class, 'addProductView'])->middleware('auth')->name('addProduct');
-    Route::post('/saveProduct', [ProductController::class, 'saveProduct'])->middleware('auth')->name('saveProduct');
+    Route::get('/addProduct', [AdminController::class, 'addProductView'])->middleware(['auth', 'role-auth'])->name('addProduct');
+    Route::post('/saveProduct', [ProductController::class, 'saveProduct'])->middleware(['auth', 'role-auth'])->name('saveProduct');
 
     // UPDATE PRODUCT
-    Route::get('/editProduct/{id}', [AdminController::class, 'editProductView'])->middleware('auth')->name('editProduct');
-    Route::post('/updateProduct/{id}', [ProductController::class, 'updateProduct'])->middleware('auth')->name('updateProduct');
+    Route::get('/editProduct/{id}', [AdminController::class, 'editProductView'])->middleware(['auth', 'role-auth'])->name('editProduct');
+    Route::post('/updateProduct/{id}', [ProductController::class, 'updateProduct'])->middleware(['auth', 'role-auth'])->name('updateProduct');
 
     // DELETE PRODUCT
-    Route::delete('/deleteProduct/{id}', [ProductController::class, 'deleteProduct'])->middleware('auth')->name('deleteProduct');
+    Route::delete('/deleteProduct/{id}', [ProductController::class, 'deleteProduct'])->middleware(['auth', 'role-auth'])->name('deleteProduct');
 
     // DASHBOARD
-    Route::get('/', [AdminController::class, 'dashboardView'])->name('dashboard')->middleware('auth');
-});
-
-Route::get('list', function () {
-    return view('listproduk');
+    Route::get('/', [AdminController::class, 'dashboardView'])->name('dashboard')->middleware(['auth', 'role-auth']);
 });
